[Feat] Changed How Cells Judge To Move Below From Checking Below Block State To Setting New Pos To Move

+++ WinAPI_Project_Puzzle_Game/include/Object/Block.h
@@ -36,8 +36,8 @@ public :
 		m_Texture = Texture;
 		if (m_Texture)
 		{
-			m_Size.x = m_Texture->GetWidth();
-			m_Size.y = m_Texture->GetHeight();
+			m_Size.x = (float)m_Texture->GetWidth();
+			m_Size.y = (float)m_Texture->GetHeight();
 		}
 	}
 	void SetMoveEnable(bool Enable)
